onkyo-api/pyproject.toml
[tool.poetry]
name = "onkyo-api"
version = "0.1.0"
description = ""
authors = ["Marcin Tyszkiewicz <56219944+mtyszkiewicz@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
onkyo-eiscp = "^1.2.7"
fastapi = "^0.110.1"
pydantic-settings = "^2.2.1"
uvicorn = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
